[Issue #98] Move Anders' steps. Minor refctoring
